+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## 0.0.3
+## 0.1.0
 
 - Add list command
+- Add install command
+- Add uninstall command
+- Add colored UI
 
 ## 0.0.2